+++ src/types.ts
@@ -1355,14 +1355,72 @@ export type ResolveMacroContext<
 
 export type ContextAppendType = 'append' | 'override'
 
-export type HigherOrderFunction<
-	T extends (...arg: unknown[]) => Function = (...arg: unknown[]) => Function
-> = (fn: T, request: Request) => ReturnType<T>
-
 // new Elysia()
 // 	.wrap((fn) => {
 // 		return fn()
 // 	})
+export type HigherOrderFunction<
+	T extends (...arg: unknown[]) => Function = (...arg: unknown[]) => Function
+> = (fn: T, request: Request) => ReturnType<T>
+
+type SetContentType =
+	| 'application/octet-stream'
+	| 'application/vnd.ms-fontobject'
+	| 'application/epub+zip'
+	| 'application/gzip'
+	| 'application/json'
+	| 'application/ld+json'
+	| 'application/ogg'
+	| 'application/pdf'
+	| 'application/rtf'
+	| 'application/wasm'
+	| 'application/xhtml+xml'
+	| 'application/xml'
+	| 'application/zip'
+	| 'text/css'
+	| 'text/csv'
+	| 'text/html'
+	| 'text/calendar'
+	| 'text/javascript'
+	| 'text/plain'
+	| 'text/xml'
+	| 'image/avif'
+	| 'image/bmp'
+	| 'image/gif'
+	| 'image/x-icon'
+	| 'image/jpeg'
+	| 'image/png'
+	| 'image/svg+xml'
+	| 'image/tiff'
+	| 'image/webp'
+	| 'multipart/mixed'
+	| 'multipart/alternative'
+	| 'multipart/form-data'
+	| 'audio/aac'
+	| 'audio/x-midi'
+	| 'audio/mpeg'
+	| 'audio/ogg'
+	| 'audio/opus'
+	| 'audio/webm'
+	| 'video/x-msvideo'
+	| 'video/quicktime'
+	| 'video/x-ms-wmv'
+	| 'video/x-msvideo'
+	| 'video/x-flv'
+	| 'video/av1'
+	| 'video/mp4'
+	| 'video/mpeg'
+	| 'video/ogg'
+	| 'video/mp2t'
+	| 'video/webm'
+	| 'video/3gpp'
+	| 'video/3gpp2'
+	| 'font/otf'
+	| 'font/ttf'
+	| 'font/woff'
+	| 'font/woff2'
+	| 'model/gltf+json'
+	| 'model/gltf-binary'
 
 export type HTTPHeaders = Record<string, string> & {
 	// Authentication
@@ -1422,7 +1480,7 @@ export type HTTPHeaders = Record<string, string> & {
 
 	// Message body information
 	'content-length'?: string
-	'content-type'?: string
+	'content-type'?: SetContentType | {}
 	'content-encoding'?: string
 	'content-language'?: string
 	'content-location'?: string