+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.14 - 23 Sep 2024
+Feature:
+- add auto-completion to `Content-Type` headers
+
+Bug fix:
+- exclude file from Bun native static response until Bun support
+- set 'text/plain' for string if no content-type is set for native static response
+
 # 1.1.13 - 18 Sep 2024
 Feature:
 - [#813](https://github.com/elysiajs/elysia/pull/813) allow UnionEnum to get readonly array by @BleedingDev